Add tariff types
