+++ api/tariff.go
@@ -1,0 +1,8 @@
+package api
+
+type TariffType int
+
+const (
+	TariffTypePrice TariffType = iota
+	TariffTypeCo2
+)
