+++ tariff/electricitymaps.go
@@ -123,3 +123,8 @@ func (t *ElectricityMaps) Rates() (api.Rates, error) {
 func (t *ElectricityMaps) IsDynamic() bool {
 	return true
 }
+
+// Type returns the tariff type
+func (t *ElectricityMaps) Type() api.TariffType {
+	return api.TariffTypeCo2
+}
